client/requirements.txt
discord.py[voice]
#git+https://github.com/Sheppsu/discord-ext-listening
python-dotenv
requests
aiohttp
pytz
psutil
sqlalchemy==2.0.7
Pillow
PyNaCl
eventlet